modifyng constructor and methods

+++ Lab-11/KenaKata/src/main/java/kenakata/catalog/AddOn.java
@@ -1,10 +1,10 @@
 package kenakata.catalog;
 
 public abstract class AddOn implements Chargeable {
-    public int unitVat() {
+    public long unitVat() {
     return 0;}
 
-    public int unitCharge() {
+    public long unitCharge() {
     return 0;}
 
 }
